requirements.txt
alembic==1.4.3
aniso8601==8.0.0
autopep8==1.5.4
bcrypt==3.2.0
cffi==1.14.3
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pyodbc==4.0.30
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.1
Werkzeug==1.0.1